requirements.txt
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
pandas>=2.0.0,<3.0.0
seaborn>=0.13.0,<1.0.0

nibabel>=5.1.0,<6.0.0
pillow>=10.0.0,<12.0.0
scikit-image>=0.21.0,<1.0.0

torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
pytorch-msssim>=1.0.0,<2.0.0

optuna>=3.4.0,<5.0.0
tqdm>=4.65.0,<5.0.0
